Atom for git commits and remove macOS setup script
The .gitconfig was updated to use Atom to author git commits.

The macOS setup scripts was removed because it must be heavily modified
after every major macOS release (things change, move, become
deprecated, etc.) and many of the settings/preferences are fairly simple
to make manually (some are even the default with never versions of
macOS).